Reduce key length metadata size to use u32
Reduces the space taken when serializing map keys by changing how key
lengths are encoded from `u64` to `u32`. This improves overall
efficiency of model file handling operations.

+++ llm-runner/src/transformer/model.rs
@@ -1200,7 +1200,7 @@ mod tests {
 
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1406,7 +1406,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1472,7 +1472,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1528,7 +1528,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1650,7 +1650,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1785,7 +1785,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1862,7 +1862,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1907,7 +1907,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
